# contracts/openapi/customer-bff-v1.yaml
openapi: 3.1.0
info:
  title: Customer Consultation BFF API
  version: 1.0.0
  description: Customer-channel consultation, feedback, retry and ticket handoff. Caller identity comes only from the bearer token.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/customer/consultations/answers:
    post:
      operationId: answerCustomerConsultation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerAnswerRequest'
      responses:
        '200':
          description: Channel-safe answer with citations and attempt identity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAnswerResponse'
        '400': { description: Invalid channel input. }
        '401': { description: Missing or invalid customer token. }
        '403': { description: Customer cannot access the requested conversation. }
        '429': { description: Per-customer consultation rate limit reached. }
        '502': { description: A downstream service failed. }
  /api/v1/customer/consultations/answers/stream:
    post:
      operationId: streamCustomerConsultation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerAnswerRequest'
      responses:
        '200':
          description: Named session, metadata, delta, heartbeat, citation, completed or error events.
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/CustomerStreamEvent'
  /api/v1/customer/consultations/{conversationId}/attempts/{attemptId}/feedback:
    put:
      operationId: recordCustomerAnswerFeedback
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/AttemptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerFeedbackRequest'
      responses:
        '204': { description: Feedback recorded for the answer attempt. }
        '403': { description: Customer cannot access the conversation. }
        '409': { description: The answer attempt is not in a feedback-eligible state. }
  /api/v1/customer/consultations/{conversationId}/attempts/{attemptId}/retry:
    post:
      operationId: retryCustomerAnswer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/AttemptId'
      responses:
        '200':
          description: A new answer attempt linked to the original attempt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAnswerResponse'
  /api/v1/customer/consultations/{conversationId}/attempts/{attemptId}/handoffs:
    post:
      operationId: handoffCustomerConsultation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/AttemptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultationHandoffRequest'
      responses:
        '202':
          description: An idempotent ticket-agent task was accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketHandoffReceipt'
        '409': { description: Handoff state or idempotency conflict. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
    AttemptId:
      name: attemptId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
  schemas:
    CustomerAnswerRequest:
      type: object
      additionalProperties: false
      required: [question]
      properties:
        conversationId:
          type: [string, 'null']
          maxLength: 64
        question:
          type: string
          minLength: 1
          maxLength: 2000
    CustomerAnswerResponse:
      type: object
      additionalProperties: false
      required: [conversationId, attemptId, answer, citations, refused, traceId]
      properties:
        conversationId: { type: string }
        attemptId: { type: string }
        retryOfAttemptId: { type: [string, 'null'] }
        answer: { type: string }
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        refused: { type: boolean }
        refusalReason: { type: [string, 'null'] }
        traceId: { type: string }
    Citation:
      type: object
      additionalProperties: false
      required: [documentId, version, sectionId, title]
      properties:
        documentId: { type: string }
        version: { type: string }
        sectionId: { type: string }
        title: { type: string }
    AnswerFeedbackRequest:
      type: object
      additionalProperties: false
      required: [rating]
      properties:
        rating:
          type: string
          enum: [HELPFUL, NOT_HELPFUL]
        reasonCode: { type: [string, 'null'], maxLength: 64 }
        comment: { type: [string, 'null'], maxLength: 500 }
    ConsultationHandoffRequest:
      type: object
      additionalProperties: false
      required: [reasonCode]
      properties:
        reasonCode: { type: string, minLength: 1, maxLength: 64 }
    TicketHandoffReceipt:
      type: object
      additionalProperties: false
      required: [taskId, status, duplicate]
      properties:
        taskId: { type: string }
        status: { type: string, enum: [ACCEPTED] }
        duplicate: { type: boolean }
    CustomerStreamEvent:
      oneOf:
        - $ref: '#/components/schemas/SessionEvent'
        - $ref: '#/components/schemas/MetadataEvent'
        - $ref: '#/components/schemas/DeltaEvent'
        - $ref: '#/components/schemas/HeartbeatEvent'
        - $ref: '#/components/schemas/CitationEvent'
        - $ref: '#/components/schemas/CompletedEvent'
        - $ref: '#/components/schemas/ErrorEvent'
    SessionEvent:
      type: object
      required: [conversationId, attemptId]
      properties:
        conversationId: { type: string }
        attemptId: { type: string }
        retryOfAttemptId: { type: [string, 'null'] }
    MetadataEvent:
      type: object
      required: [traceId]
      properties:
        traceId: { type: string }
    DeltaEvent:
      type: object
      required: [text]
      properties:
        text: { type: string }
    HeartbeatEvent:
      type: object
      required: [epochMillis]
      properties:
        epochMillis: { type: integer, format: int64 }
    CitationEvent:
      type: object
      required: [citation]
      properties:
        citation:
          $ref: '#/components/schemas/Citation'
    CompletedEvent:
      type: object
      additionalProperties: false
    ErrorEvent:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }